adicionando início dos testes

+++ src/test/java/br/ufpe/cin/spgroup/dslcotasgen/dslcotasgen/QuadroVaga.java
@@ -3,16 +3,16 @@ package br.ufpe.cin.spgroup.dslcotasgen.dslcotasgen;
 public class QuadroVaga {
 
 	private String sigla;
-	private Long quantidade;
-	public QuadroVaga(String sigla, Long quantidade) {
+	private Integer quantidade;
+	public QuadroVaga(String sigla, Integer quantidade) {
 		super();
 		this.sigla = sigla;
 		this.quantidade = quantidade;
 	}
-	public Long getQuantidade() {
+	public Integer getQuantidade() {
 		return quantidade;
 	}
-	public void setQuantidade(Long quantidade) {
+	public void setQuantidade(Integer quantidade) {
 		this.quantidade = quantidade;
 	}
 	public String getSigla() {
